src/requirements.txt
Flask==1.0.2
requests==2.20.1
google-cloud-storage
pymysql